requirements.txt
python-telegram-bot==21.6
requests
python-dotenv
openai
supabase
requests